+++ src/components/landing/index.jsx
@@ -14,7 +14,8 @@ import Help1 from "../../assets/landing/help 1.png";
 import Help2 from "../../assets/landing/help 2.png";
 import Help3 from "../../assets/landing/help 3.png";
 import Help4 from "../../assets/landing/help 4.png";
-import WalletsPng from "../../assets/landing/wallets.png";
+import WalletsPng from "../../assets/landing/Wallets.png";
+import MobileWalletPng from "../../assets/landing/WalletImage-2.png";
 import Benefit1 from "../../assets/landing/benefit1.png";
 import Benefit2 from "../../assets/landing/benefit2.png";
 import ServiceLine1 from "../../assets/landing/service line left.svg";
@@ -378,14 +379,14 @@ export const Benefits = () => {
         subtitle={`Nefentus prioritizes your customers' convenience at every step.`}
       />
       <div className="layout-benefits">
-        <div style={{ borderRadius: "16px" }}>
+        <div style={{ borderRadius: "16px" }} className="web-outlook-options">
           <img
             src={WalletsPng}
             style={{
               height: "100%",
               width: "100%",
               objectFit: "cover",
-              objectPosition: "center top",
+              objectPosition: "center ",
             }}
             alt="Wallets for cryptocurrency"
           />
@@ -395,6 +396,17 @@ export const Benefits = () => {
               Customers have the flexibility to pay using their preferred wallet
               & currency, spanning across Bitcoin, Ethereum and numerous other
               networks.
+            </p>
+          </div>
+        </div>
+        <div className="mobile-outlook-img">
+          <div className="mobile-view-wallet-images">
+            <p className="crypto-wallet-mobile-title mobile-subtext">
+              Hundreds of currencies, all Web3 wallets
+            </p>
+            <p className="sub-title crypto-wallet-mobile-subtext">
+              Customers can pay with their preferred wallet and currency across
+              Bitcoin, Ethereum, and other networks
             </p>
           </div>
         </div>